refactor: Move `FuzzySelectScript` to fzf package
